LayoutTests/inspector/debugger/dynamic-script-tag-expected.txt

Tests that inline scripts and document.write scripts get different uiSourceCodes with different URLs. Bug 87119

Debugger was enabled.
Dumping scripts:
script1: LayoutTests/inspector/debugger/dynamic-script-tag.html
Is anonymous:false
Is inline script:false
script2: LayoutTests/inspector/debugger/dynamic-script-tag.html
Is anonymous:false
Is inline script:false
Dumping uiSourceCodes:
UISourceCode: LayoutTests/inspector/debugger/dynamic-script-tag.html
UISourceCode is editable: false
UISourceCode is content script: false
Mime type: text/javascript
UISourceCode content: function foo() { }
UISourceCode: LayoutTests/inspector/debugger/dynamic-script-tag.html
UISourceCode is editable: false
UISourceCode is content script: false
Mime type: text/javascript
UISourceCode content: function bar() { }
Debugger was disabled.
